+++ data-source/json/test_list_collector_section_summary.json
@@ -438,7 +438,8 @@
                         },
                         {
                             "id": "people-list-section-summary",
-                            "type": "SectionSummary"
+                            "title": "List",
+                            "type": "ListCollectorSummary"
                         }
                     ]
                 }
